Add encryptedSecret to User
- Change client UserEntity
- Change server UserEntity
- Change client UserRepository
- Change server UserRepository
- Change userSchema
- Add index.js for models
- Move sequelize Models to index.js
Change userController to support encryptedSecret
