+++ client/src/interface-adapters/repositories/UserRepository.js
@@ -16,6 +16,9 @@ export const UserRepository = class UserRepository {
           username: userData.username,
           email: userData.email,
           password: userData.password,
+          encryptedSecret: userData.encryptedSecret,
+          encryptionIv: userData.encryptionIv,
+          encryptionSalt: userData.encryptionSalt,
           isNewsletterAllowed: userData.isNewsletterAllowed,
         }),
         credentials: "include",
