+++ server/src/entities/UserEntity.js
@@ -34,6 +34,9 @@ exports.UserEntity = class UserEntity {
       "username",
       "email",
       "password",
+      "encryptedSecret",
+      "encryptionIv",
+      "encryptionSalt",
       "isVerified",
       "isBlocked",
       "isNewsletterAllowed",
