+++ server/src/interface-adapters/repositories/userRepository.js
@@ -1,78 +1,4 @@
-const { Sequelize, Model, DataTypes } = require("sequelize");
-const sequelize = new Sequelize(process.env.DATABASE_URL, {
-  logging: false,
-  dialectOptions: {
-    useUTC: true, // for reading from database
-  },
-  timezone: "+00:00", // for writing to database
-});
-
-class User extends Model {}
-
-User.init(
-  {
-    userId: {
-      type: DataTypes.INTEGER,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    username: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-    },
-    email: {
-      type: DataTypes.STRING,
-      allowNull: false,
-      unique: true,
-    },
-    password: {
-      type: DataTypes.STRING,
-      allowNull: false,
-    },
-    passwordUpdatedAt: {
-      type: DataTypes.DATE,
-      defaultValue: Sequelize.literal("CURRENT_TIMESTAMP"),
-    },
-    isVerified: {
-      type: DataTypes.BOOLEAN,
-      defaultValue: false,
-    },
-    isBlocked: {
-      type: DataTypes.BOOLEAN,
-      defaultValue: false,
-    },
-    isNewsletterAllowed: {
-      type: DataTypes.BOOLEAN,
-      defaultValue: false,
-    },
-    verificationCode: {
-      type: DataTypes.STRING,
-    },
-    verificationAttempts: {
-      type: DataTypes.INTEGER,
-      defaultValue: 0,
-    },
-    lastVerificationAttempt: {
-      type: DataTypes.DATE,
-      defaultValue: null,
-    },
-    createdAt: {
-      type: DataTypes.DATE,
-      defaultValue: Sequelize.literal("CURRENT_TIMESTAMP"),
-    },
-    updatedAt: {
-      type: DataTypes.DATE,
-      defaultValue: null,
-    },
-  },
-  {
-    sequelize,
-    modelName: "user",
-    timestamps: false,
-    underscored: true,
-  }
-);
+const { User } = require("../models");
 
 exports.UserRepository = class UserRepository {
   async existsByUsername(username) {
